subscribe consumers to specific partition-topic combinations

+++ server/src/hash.rs
@@ -2,7 +2,7 @@ use std::ptr::hash;
 
 pub struct Hasher {}
 impl Hasher {
-    pub fn hash(n: u32) -> String {
+    pub fn hash(n: u16) -> String {
         let digest = md5::compute(format!("{}", n));
         format!("{:x}", digest)
     }
